+++ test/unit/router.spec.js
@@ -1,9 +1,9 @@
-var expect          = require('chai').expect;
-var proxyTable      = require('./_libs').proxyTable;
+var expect = require('chai').expect;
+var router = require('./_libs').router;
 
-describe('proxyTable unit test', function() {
+describe('router unit test', function() {
 
-    describe('proxyTable.createProxyOptions', function() {
+    describe('router.createProxyOptions', function() {
         var req, config, result;
 
         beforeEach(function() {
@@ -22,7 +22,7 @@ describe('proxyTable unit test', function() {
             configProxyTable = {
                 target: 'http://localhost:6000',
                 changeOrigin: true,                // other options should be returned, such as changeOrigin
-                proxyTable: {
+                router: {
                     'alpha.localhost': 'http://localhost:6001',
                     'beta.localhost': 'http://localhost:6002',
                     'gamma.localhost/api': 'http://localhost:6003',
@@ -34,9 +34,9 @@ describe('proxyTable unit test', function() {
             };
         });
 
-        describe('without proxyTable config', function() {
-            it('should return the normal target when proxyTable not present in config', function() {
-                result = proxyTable.createProxyOptions(req, config);
+        describe('without router config', function() {
+            it('should return the normal target when router not present in config', function() {
+                result = router.createProxyOptions(req, config);
                 expect(result.target).to.equal('http://localhost:6000');
                 expect(result).not.to.equal(config);            // should return cloned object
                 expect(result).to.deep.equal(config);           // clone content should match
@@ -44,17 +44,17 @@ describe('proxyTable unit test', function() {
             });
         });
 
-        describe('with just the host in proxyTable config', function() {
-            it('should target http://localhost:6001 when for proxyTable:"alpha.localhost"', function() {
+        describe('with just the host in router config', function() {
+            it('should target http://localhost:6001 when for router:"alpha.localhost"', function() {
                 req.headers.host = 'alpha.localhost';
-                result = proxyTable.createProxyOptions(req, configProxyTable);
+                result = router.createProxyOptions(req, configProxyTable);
                 expect(result.target).to.equal('http://localhost:6001');
                 expect(result.changeOrigin).to.be.true;
             });
 
-            it('should target http://localhost:6002 when for proxyTable:"beta.localhost"', function() {
+            it('should target http://localhost:6002 when for router:"beta.localhost"', function() {
                 req.headers.host = 'beta.localhost';
-                result = proxyTable.createProxyOptions(req, configProxyTable);
+                result = router.createProxyOptions(req, configProxyTable);
                 expect(result.target).to.equal('http://localhost:6002');
                 expect(result.changeOrigin).to.be.true;
             });
@@ -63,7 +63,7 @@ describe('proxyTable unit test', function() {
         describe('with host and host + path config', function() {
             it('should target http://localhost:6004 without path', function() {
                 req.headers.host = 'gamma.localhost';
-                result = proxyTable.createProxyOptions(req, configProxyTable);
+                result = router.createProxyOptions(req, configProxyTable);
                 expect(result.target).to.equal('http://localhost:6004');
                 expect(result.changeOrigin).to.be.true;
             });
@@ -71,7 +71,7 @@ describe('proxyTable unit test', function() {
             it('should target http://localhost:6003 exact path match', function() {
                 req.headers.host = 'gamma.localhost';
                 req.url = '/api';
-                result = proxyTable.createProxyOptions(req, configProxyTable);
+                result = router.createProxyOptions(req, configProxyTable);
                 expect(result.target).to.equal('http://localhost:6003');
                 expect(result.changeOrigin).to.be.true;
             });
@@ -79,39 +79,39 @@ describe('proxyTable unit test', function() {
             it('should target http://localhost:6004 when contains path', function() {
                 req.headers.host = 'gamma.localhost';
                 req.url = '/api/books/123';
-                result = proxyTable.createProxyOptions(req, configProxyTable);
+                result = router.createProxyOptions(req, configProxyTable);
                 expect(result.target).to.equal('http://localhost:6003');
                 expect(result.changeOrigin).to.be.true;
             });
         });
 
         describe('with just the path', function() {
-            it('should target http://localhost:6005 with just a path as proxyTable config', function() {
+            it('should target http://localhost:6005 with just a path as router config', function() {
                 req.url = '/rest';
-                result = proxyTable.createProxyOptions(req, configProxyTable);
+                result = router.createProxyOptions(req, configProxyTable);
                 expect(result.target).to.equal('http://localhost:6005');
                 expect(result.changeOrigin).to.be.true;
             });
 
-            it('should target http://localhost:6005 with just a path as proxyTable config', function() {
+            it('should target http://localhost:6005 with just a path as router config', function() {
                 req.url = '/rest/deep/path';
-                result = proxyTable.createProxyOptions(req, configProxyTable);
+                result = router.createProxyOptions(req, configProxyTable);
                 expect(result.target).to.equal('http://localhost:6005');
                 expect(result.changeOrigin).to.be.true;
             });
 
-            it('should target http://localhost:6000 path in not present in proxyTable config', function() {
+            it('should target http://localhost:6000 path in not present in router config', function() {
                 req.url = '/unknow-path';
-                result = proxyTable.createProxyOptions(req, configProxyTable);
+                result = router.createProxyOptions(req, configProxyTable);
                 expect(result.target).to.equal('http://localhost:6000');
                 expect(result.changeOrigin).to.be.true;
             });
         });
 
-        describe('matching order of proxyTable config', function() {
+        describe('matching order of router config', function() {
             it('should return first matching target when similar paths are configured', function() {
                 req.url = '/some/specific/path';
-                result = proxyTable.createProxyOptions(req, configProxyTable);
+                result = router.createProxyOptions(req, configProxyTable);
                 expect(result.target).to.equal('http://localhost:6006');
                 expect(result.changeOrigin).to.be.true;
             });