+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [develop](https://github.com/chimurai/http-proxy-middleware/releases/tag/v0.16.0)
+- feat(router): deprecate and renamed `proxyTable` option to `router`
+
 ## [v0.15.2](https://github.com/chimurai/http-proxy-middleware/releases/tag/v0.15.2)
 - fix(websocket): fixes websocket upgrade